Bug fixes and Settings page updates
Fix a bug where the app would improperly reload user data when "afterConfirm" or "noTimer" settings were on.

Locked settings activity orientation to portrait and gave it a tile "Settings".

Removed the option for "daily notification" setting. This option was not previously implemented. I removed the option because I am unsure if users would like this option and at this time do not find it necessary.